requirements.txt
streamlit
google-generativeai
python-dotenv
audio-recorder-streamlit
gTTS
